refactor: migrate custom input to shared components
