+++ src/Common/index.ts
@@ -60,7 +60,6 @@ export * from './RJSF'
 export * from './DevtronProgressing'
 export { default as ChartVersionAndTypeSelector } from './ChartVersionAndTypeSelector'
 export * from './AddCDButton'
-export * from './CustomInput'
 export * from './DraggableWrapper'
 export * from './Pagination'
 export * from './Markdown'
