+++ src/Shared/Components/index.ts
@@ -76,3 +76,4 @@ export * from './ThemeSwitcher'
 export * from './TargetPlatforms'
 export * from './UnsavedChanges'
 export * from './UnsavedChangesDialog'
+export * from './CustomInput'
